backend/requirements.txt
Django>=3.2,<4.0
psycopg2-binary
djangorestframework
djangorestframework-simplejwt  # For JWT tokens
social-auth-app-django  # For OAuth
django-cors-headers  # For security
django-oauth-toolkit 
python-dotenv
requests
django-otp==1.2.2
qrcode==7.4.2
pyotp==2.9.0
pillow==10.0.0


daphne
channels
channels_redis